requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==0.8.11
PyMuPDF==1.23.8
openai>=1.10.0
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.4.2
pydantic-settings==2.0.3
